feat: Add hash set implementation which can be stored as account
We want to use hash sets as queues for nullifiers and address spaces.
However, the `std` and `hashbrown` implementations are dynamically
allocated and resized, which doesn't serve our purpose of storing the
underlying data as Solana account.

Therefore, we implement our own hash set with the following properties:

* [Quadratic probing](https://en.wikipedia.org/wiki/Quadratic_probing)
  as a solution for hash collisions.
* Load factor 0.7 (70%) - which means that the size of the structure
  always includes at least 70% of additional space (more than expected
  elements) for efficiency.
* The final size is a prime number greather than the size of elements
  and size determined by low factor. To find the prime number, we probe
  6k + 1 and 6k + 5 candidates. We check whether they are prime in a
  brute-force way, because we want to be sure there are no false
  positives.

+++ programs/account-compression/src/instructions/append_leaves.rs
@@ -121,10 +121,10 @@ pub mod sdk {
                 index: 1u64,
                 owner: payer,
                 delegate: None,
-                height: STATE_MERKLE_TREE_HEIGHT as u64,
-                changelog_size: STATE_MERKLE_TREE_CHANGELOG as u64,
-                roots_size: STATE_MERKLE_TREE_ROOTS as u64,
-                canopy_depth: STATE_MERKLE_TREE_CANOPY_DEPTH as u64,
+                height: STATE_MERKLE_TREE_HEIGHT,
+                changelog_size: STATE_MERKLE_TREE_CHANGELOG,
+                roots_size: STATE_MERKLE_TREE_ROOTS,
+                canopy_depth: STATE_MERKLE_TREE_CANOPY_DEPTH,
                 associated_queue,
             };
         Instruction {
